authentication/requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
pydantic
pydantic_settings
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
redis